requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
spacy==3.8.13
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

sentence-transformers==2.7.0
torch==2.2.2
scikit-learn==1.6.1
xgboost==2.1.4
pandas==2.2.3
numpy==2.2.5
scipy==1.15.3
pyspellchecker==0.8.2